--- src/main/java/chzzk/grassdiary/auth/config/WebConfig.java
@@ -1,41 +1,0 @@
-package chzzk.grassdiary.auth.config;
-
-import chzzk.grassdiary.auth.common.AuthMemberResolver;
-import java.util.List;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-@RequiredArgsConstructor
-public class WebConfig implements WebMvcConfigurer {
-
-    private final AuthInterceptor authInterceptor;
-    private final AuthMemberResolver authMemberResolver;
-
-    @Override
-    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
-        resolvers.add(authMemberResolver);
-    }
-
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(authInterceptor)
-                .order(1)
-                .excludePathPatterns("/css/**", "/*.ico", "/error")
-                .excludePathPatterns("/api/auth/**/google")
-                .addPathPatterns("/api/**");
-    }
-
-    @Override
-    public void addCorsMappings(CorsRegistry registry) {
-        registry.addMapping("/**")
-                .allowedOrigins("http://localhost:3000")
-                .allowedMethods("*")
-                .allowedHeaders("*")
-                .allowCredentials(true);
-    }
-}
