+++ src/main/java/chzzk/grassdiary/config/WebMvcConfig.java
@@ -1,13 +1,36 @@
 package chzzk.grassdiary.config;
 
+import chzzk.grassdiary.auth.common.AuthMemberResolver;
+import chzzk.grassdiary.auth.config.AuthInterceptor;
+import java.util.List;
+import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @Configuration
 @EnableWebMvc
+@RequiredArgsConstructor
 public class WebMvcConfig implements WebMvcConfigurer {
+    private final AuthInterceptor authInterceptor;
+    private final AuthMemberResolver authMemberResolver;
+
+    @Override
+    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
+        resolvers.add(authMemberResolver);
+    }
+
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(authInterceptor)
+                .order(1)
+                .excludePathPatterns("/css/**", "/*.ico", "/error")
+                .excludePathPatterns("/api/auth/**/google")
+                .addPathPatterns("/api/**");
+    }
 
     @Override
     public void addCorsMappings(CorsRegistry registry) {
